Refactor imports to use models_public
