+++ api-rework/invasives/api/models/activity/biocontrol/biocontrol_agent_count.py
@@ -2,7 +2,7 @@ from django.core.validators import MinValueValidator
 from django.db import models
 
 from api.models.activity.abstract_sub_tables import BaseOneToManyActivityTable
-from api.models.codes import (
+from api.models_public.codes import (
     BiocontrolAgentCode,
     BioAgentLifeStageCode,
     TerrestrialPlantCode,
