+++ api-rework/invasives/api/models/activity/biocontrol/spread_results.py
@@ -1,7 +1,7 @@
 from django.db import models
 from django.core.validators import MaxValueValidator
 from api.models.activity.abstract_sub_tables import BaseOneToOneActivityTable
-from api.models.codes import MesoslopePositionCode, SiteSurfaceShapeCode
+from api.models_public.codes import MesoslopePositionCode, SiteSurfaceShapeCode
 
 
 class SpreadResults(BaseOneToOneActivityTable):
