--- api-rework/invasives/api/models/activity/observations/__init__.py
@@ -1,12 +1,0 @@
-from .aquatic_plant_observation_detail import *
-from .pre_treatment_observation import *
-from .shoreline_types import *
-from .suitable_for_biocontrol import *
-from .terrestrial_plant_observation_detail import *
-from .terrestrial_plant_observation_info import *
-from .voucher_specimen import *
-from .waterbody_data import *
-from .waterbody_flow_detail import *
-from .waterbody_level_management import *
-from .waterbody_substrate_type import *
-from .waterbody_use import *
