+++ api-rework/invasives/api/models/activity/observations/terrestrial_plant_observation_detail.py
@@ -1,13 +1,13 @@
 from django.db import models
 from django.core.exceptions import ValidationError
 from api.models.activity.abstract_sub_tables import BaseOneToManyActivityTable
-from api.models.codes import (
+from api.models_public.codes import (
     DensityCode,
     DistributionCode,
     TerrestrialPlantCode,
     PlantLifeStageCode,
 )
-from api.models.enums import ObservationType
+from api.models_public.enums import ObservationType
 
 
 class TerrestrialPlantObservationDetail(BaseOneToManyActivityTable):
