+++ api-rework/invasives/api/models/activity/observations/terrestrial_plant_observation_info.py
@@ -1,8 +1,8 @@
 from django.db import models
 from django.core.exceptions import ValidationError
 from api.models.activity.abstract_sub_tables import BaseOneToOneActivityTable
-from api.models.codes import SlopePercentCode, AspectCode, SoilTextureCode
-from api.models.enums import YesNoUnknown
+from api.models_public.codes import SlopePercentCode, AspectCode, SoilTextureCode
+from api.models_public.enums import YesNoUnknown
 
 
 class TerrestrialPlantObservationInfo(BaseOneToOneActivityTable):
