+++ api-rework/invasives/api/models/activity/observations/voucher_specimen.py
@@ -1,6 +1,6 @@
 from django.db import models
 from api.models.activity.abstract_sub_tables import BaseOneToOneActivityTable
-from api.models.codes import TerrestrialPlantCode, AquaticPlantCode
+from api.models_public.codes import TerrestrialPlantCode, AquaticPlantCode
 
 
 class VoucherSpecimen(BaseOneToOneActivityTable):
