+++ api-rework/invasives/api/models/activity/observations/waterbody_data.py
@@ -1,7 +1,7 @@
 from django.db import models
-from api.models.enums import YesNoUnknown
+from api.models_public.enums import YesNoUnknown
 from api.models.activity.abstract_sub_tables import BaseOneToOneActivityTable
-from api.models.enums import WaterbodyType
+from api.models_public.enums import WaterbodyType
 
 
 class WaterbodyData(BaseOneToOneActivityTable):
