+++ api-rework/invasives/api/models/activity/observations/waterbody_substrate_type.py
@@ -1,5 +1,5 @@
 from django.db import models
-from api.models.codes import SubstrateCode
+from api.models_public.codes import SubstrateCode
 from api.models.activity.abstract_sub_tables import BaseOneToManyActivityTable
