+++ api-rework/invasives/api/viewsets/code.py
@@ -6,7 +6,7 @@ from rest_framework.response import Response
 from rest_framework.status import HTTP_200_OK
 from rest_framework.viewsets import ViewSet
 
-from api.models.codes import *
+from api.models_public.codes import *
 from api.serializers import CodeSerializer
